refactor: index.html에서 전체 레이아웃이 작성되어 있는 형태로 수정
