+++ src/components/header/header.js
@@ -3,14 +3,12 @@ import './Header.css';
 
 export default class Header {
   constructor({ $target }) {
-    this.$header = document.createElement('header');
-    this.$header.id = 'header';
-    $target.appendChild(this.$header);
+    this.$target = $target;
     
     this.render();
   }
 
   render() {
-    this.$header.innerHTML = html;
+    this.$target.innerHTML = html;
   }
 }
