+++ src/index.html
@@ -3,6 +3,15 @@
     <title>럭키?! 언럭키!?</title>
   </head>
   <body>
-    <div id="app"></div>
+    <div id="app">
+      <header id="header"></header>
+      <main id="main">
+        <section class="item-section"></section>
+        <div class="roullet-container">
+          <section class="roullet-section"></section>
+          <section class="roullet-history-section"></section>
+        </div>
+      </main>
+    </div>
   </body>
 </html>
